Move trait impl from committee session domain to api
